// frontend/src/app/models/addUser.ts
export class AddUser {
    _id: String='';
    rmid : String ='';
    name: String='';
    email: String='';
    password: String='';
    mobile_number : String='';
    designation : String = '';
    city : String = '';
    state : String = '';
    reportingTo : String = '';
    status: Boolean = false;
    createdon: String = '';
}